feat(ext): job info send on background using messaging
